replaced realloc by hashtable in kernel JIT compile

+++ includes/hashtable.h
@@ -10,15 +10,15 @@
 # include <stdlib.h>
 # include <limits.h>
 
-typedef struct  s_entry
+typedef struct	s_entry
 {
-    char        *key;
-    void        *value;
+	char		*key;
+	void		*value;
 }               t_entry;
 typedef struct  s_hashtable
 {
-    int         size;
-    t_entry     **table;
+	size_t		size;
+	t_entry		**table;
 }               t_hashtable;
 t_hashtable     *create_hash_table(size_t size);
 void            ht_set(t_hashtable *table, char *key, void *value);
